# PR: Improve text, code, formatting and performance and fix various issues