--- a/README.md
+++ b/README.md
@@ -1,4 +1,4 @@
-# Spyder-Workshop
+# Spyder Workshop
 
 This workshop explores some of the Spyder Python IDE's core functionality for scientific programming.
 We will work on data visualization, analysis and prediction using packages like Pandas, Matplotlib and Scikit-learn over a dataset with historical weather observations from 2006 to 2016.
@@ -6,7 +6,7 @@ We will work on data visualization, analysis and prediction using packages like 
 
 ## Prerequisites
 
-To use this tutorial, you will need to have [Spyder](https://www.spyder-ide.org) installed in a Python environment that contains at least the Numpy, Matplotlib, Pandas and Scikit-Learn packages.
+To follow along with this workshop, you will need to have [Spyder](https://www.spyder-ide.org) installed in a Python environment that contains at least the Numpy, Matplotlib, Pandas and Scikit-Learn packages.
 We recommend you [download the Anaconda Python distribution](https://www.anaconda.com/products/individual) which contains all these packages and more all in one easy-to-install place.
 
 You will also need to [download Git](https://git-scm.com/downloads), the popular version control tool, to work with the contents of this repository.
@@ -76,7 +76,8 @@ Use the Variable Explorer to verify that our data is not ordered by default.
 7. First, parse the date and create a new DataFrame with our data ordered by it:
 
 ```python
-weather_data['Formatted Date'] = pd.to_datetime(weather_data['Formatted Date'])
+weather_data['Formatted Date'] = pd.to_datetime(
+    weather_data['Formatted Date'], format='%Y-%m-%s %H:%M:%S %z')
 weather_data_ordered = weather_data.sort_values(by='Formatted Date')
 ```
 
@@ -87,22 +88,23 @@ Notice in the Variable Explorer that the DataFrame's index (the `Index` column o
 Reset the index to be in order again:
 
 ```python
-weather_data_ordered = weather_data_ordered.reset_index(drop=True)
+weather_data_ordered.reset_index(drop=True, inplace=True)
 ```
 
 We also see that there are some qualitative variables, which can make our analysis more difficult.
 For this reason, we want to filter out these columns, and stick to the ones that give us numerical information:
 
 ```python
-weather_data_ordered = weather_data_ordered.drop(
-    columns=['Summary', 'Precip Type', 'Loud Cover', 'Daily Summary'])
+weather_data_ordered.drop(
+    columns=['Summary', 'Precip Type', 'Loud Cover', 'Daily Summary'],
+    inplace=True)
 ```
 
 9. Plot the `Temperature (C)` column versus the `Date` column to see how temperature changes over time:
 
 ```python
 weather_data_ordered.plot(
-    x='Formatted Date', y=['Temperature (C)'], color='red', figsize=(15, 8))
+    x='Formatted Date', y='Temperature (C)', color='red', figsize=(15, 8))
 ```
 
 10. Switch to Spyder's Plots pane, in the same top-right section of the interface as the Variable Explorer, to view your figure.
@@ -112,7 +114,9 @@ weather_data_ordered.plot(
 12. Plot both the `Temperature (C)` and `Humidity` columns versus the `Date` column to examine how both variables change over the years:
 
 ```python
-weather_data_ordered.plot(subplots=True, x= 'Formatted Date', y= ['Temperature (C)', 'Humidity'],figsize=(15, 8))
+weather_data_ordered.plot(
+    subplots=True, x='Formatted Date', y=['Temperature (C)', 'Humidity'],
+    figsize=(15, 8))
 ```
 
 13. Explore your data!
@@ -128,13 +132,14 @@ We have written a function for you in the `utils.py` file that creates a new col
 14. Import the function from the `utils` module, so you can use it in your script:
 
 ```python
-from utils import (aggregate_by_year)
+from utils import aggregate_by_year
 ```
 
 15. Then, use it to aggregate by year and plot the data:
 
- ```python
- weather_data_by_year = aggregate_by_year(weather_data_ordered, 'Formatted Date')
+```python
+weather_data_by_year = aggregate_by_year(
+    weather_data_ordered, 'Formatted Date')
 ```
 
 16. Try writing your own function that averages the weather data by month and plots it.
@@ -148,7 +153,7 @@ For this, we have written another function in `utils.py`.
 17. First, import the new function:
 
 ```python
-from utils import (aggregate_by_year, plot_correlations)
+from utils import aggregate_by_year, plot_correlations
 ```
 
 18. Now, use it to plot the correlations between the variables:
@@ -162,14 +167,14 @@ Like before, open the Plots pane to view the correlations plot.
 19. Now, import the `plot_color_gradients()` function, which will help you plot the colormap gradient to be able to interpret your correlation plot:
 
 ```python
-from utils import (aggregate_by_year, plot_correlations, plot_color_gradients)
+from utils import aggregate_by_year, plot_correlations, plot_color_gradients
 ```
 
 20. Plot the colormap gradient using the function you imported:
 
 ```python
-cmap_category, cmap_list = ('Plot gradients convention', ['viridis', ])
-plot_color_gradients(cmap_category, cmap_list)
+plot_color_gradients(
+    cmap_category='Plot gradients convention', cmap_list=['viridis', ])
 ```
 
 21. Now, calculate the (Pearson) correlations between the different variables in our data set:
@@ -178,7 +183,7 @@ plot_color_gradients(cmap_category, cmap_list)
 weather_correlations = weather_data_ordered.corr()
 ```
 
-22. Open the object  `weather_correlations` in the Variable Explorer to see the results.
+22. Open the object `weather_correlations` in the Variable Explorer to see the results.
 
 23. You can also get just the correlation between two particular variables, such as temperature and humidity:
 
@@ -201,8 +206,8 @@ We are going to use Scikit-Learn, which is a Python package that contains tools 
 25. We will use two functions for this task which need to be imported first:
 
 ```python
-from sklearn.model_selection import train_test_split
 from sklearn import linear_model
+from sklearn.model_selection import train_test_split
 ```
 
 26. A classic way to make a predictive model is to subdivide the total set of data into two: training and test.
@@ -210,7 +215,7 @@ The training data will help us to train our model, while the test data will play
 Scikit-Learn contains a built-in function to split your data:
 
 ```python
-X_train, X_test, Y_train, Y_test = train_test_split(
+x_train, x_test, y_train, y_test = train_test_split(
     weather_data_ordered['Humidity'], weather_data_ordered['Temperature (C)'],
     test_size=0.25)
 ```
@@ -219,8 +224,8 @@ X_train, X_test, Y_train, Y_test = train_test_split(
 Create the model, then fit it with the weather data:
 
 ```python
-regresion = linear_model.LinearRegression()
-regresion.fit(X_train.values.reshape(-1, 1), Y_train.values.reshape(-1, 1))
+regression = linear_model.LinearRegression()
+regression.fit(x_train.values.reshape(-1, 1), y_train.values.reshape(-1, 1))
 ```
 
 28. To get the function's documentation in Spyder's Help Pane, place the text cursor over `LinearRegression()` and press the `Inspect` shortcut (`Ctrl+I` by default on Windows/Linux, or `Cmd-I` on macOS) .
@@ -228,14 +233,7 @@ regresion.fit(X_train.values.reshape(-1, 1), Y_train.values.reshape(-1, 1))
 29. Finally, print the coefficients of our regression:
 
 ```python
-print(regresion.intercept_, regresion.coef_)
-```
-
-and save them in variables so we can use them later:
-
-```python
-beta_0 = regresion.intercept_[0]
-beta_1 = regresion.coef_[0, 0]
+print(regression.intercept_, regression.coef_)  # beta_0, beta_1
 ```
 
 Note that this means our model is a linear function `$$y = beta_0 + beta_1 \times x$$`, where future temperature is a function of past temperature.
@@ -246,17 +244,16 @@ Note that this means our model is a linear function `$$y = beta_0 + beta_1 \time
 30. First, we want to plot our model predictions vs. our test data to see how accurate our prediction was:
 
 ```python
-Y_predict = predicted_temperature(X_test, beta_0, beta_1)
-plt.scatter(X_test, Y_test, c='red', label='observation', s=1)
-plt.scatter(X_test, Y_predict, c='blue', label='model')
-
+y_predict = regression.predict(x_test.values.reshape(-1, 1))
+plt.scatter(x_test, y_test, c='red', label='observation', s=1)
+plt.scatter(x_test, y_predict, c='blue', label='model')
 plt.xlabel('Humidity')
 plt.ylabel('Temperature (C)')
 plt.legend()
 plt.show()
 ```
 
-31. Using the coefficients found in our model, predict the temperature for a given level of humidity using the `predicted_temperature` function available in 'utils.py'.
+31. Using our model, predict the temperature for a given level of humidity.
 
 32. Finally, we want to evaluate our model performance.
 For this we will use the `explained_variance_score` metric available in `sklearn.metrics`.
@@ -270,6 +267,5 @@ from sklearn.metrics import explained_variance_score
 33. Finally, calculate the explained variance score and print it:
 
 ```python
-ev = explained_variance_score(Y_test, Y_predict)
-print(ev)
+print(explained_variance_score(y_test, y_predict))
 ```