--- a/utils.py
+++ b/utils.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-
-#
-# Copyright © Spyder Project Contributors
-# Licensed under the terms of the MIT License
 """Workshop utility functions."""
 
 # Third-party imports
@@ -13,50 +10,82 @@ def plot_correlations(df, size=10):
     """
     Plot a graphical correlation matrix for each pair of columns.
 
-    Input:
-        df: pandas DataFrame
-        size: vertical and horizontal size of the plot
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        DataFrame on which to calculate correlations.
+    size : numeric, optional
+        Vertical and horizontal size of the plot. The default is 10.
+
+    Returns
+    -------
+    fig : matplotlib.figure.Figure
+        Figure object of the generated plot.
+    ax : matplotlib.axes.Axes
+        Axes object of the generated plot.
+
     """
     corr = df.corr()
     fig, ax = plt.subplots(figsize=(size, size))
     ax.matshow(corr)
-    plt.xticks(range(len(corr.columns)), corr.columns)
-    plt.yticks(range(len(corr.columns)), corr.columns)
+    ax.set(xticks=range(len(corr.columns)), xticklabels=corr.columns)
+    ax.set(yticks=range(len(corr.columns)), yticklabels=corr.columns)
+
     plt.show()
+    return fig, ax
 
 
 def aggregate_by_year(df, date_column, figsize=(15, 8)):
     """
-    Return a DataFrame aggregate by year and plot it.
+    Aggregate a DataFrame by year and plot it.
 
-    Input:
-        df: pandas DataFrame
-        date_column: label of the column with date values
-    Return:
-        df_yearly: DataFrame grouped by month
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        DataFrame to aggregate by year and plot.
+    date_column : str
+        Label of the column containing the date values, as pd.Timestamps.
+    figsize : Tuple of (int, int), optional
+        Tuple with the size of the figure to generate in (width, height).
+        The default is (15, 8).
+
+    Returns
+    -------
+    df_yearly : pandas.DataFrameGroupBy
+        Grouped dataframe by year.
+
     """
     df['year'] = df[date_column].apply(lambda x: x.year)
     df_yearly = df.groupby('year').mean()
     df_yearly.plot(subplots=True, figsize=figsize)
-
     return df_yearly
-
-def predicted_temperature(humidity, beta_0, beta_1):
-    """
-    Return a predicted temperature based on the humidity.
-
-    Uses a linear regression as model: temperature = beta_0 + beta_1 * humidity
-    """
-    return beta_0 + beta_1 * humidity
 
 
 def plot_color_gradients(cmap_category, cmap_list):
-    """Plot a convention for color gradients used for correlations."""
+    """
+    Plot a convention for color gradients used for correlations.
+
+    Parameters
+    ----------
+    cmap_category : str
+        Category of the color map, to use for the plot title.
+    cmap_list : list of str
+        List of colormap names to plot.
+
+    Returns
+    -------
+    fig : matplotlib.figure.Figure
+        Figure object of the generated plot.
+    ax : matplotlib.axes.Axes
+        Axes object of the generated plot.
+
+    """
     gradient = np.linspace(0, 1, 256)
     gradient = np.vstack((gradient, gradient))
+
     # Create figure and adjust figure height to number of colormaps
     nrows = len(cmap_list)
-    figh = 0.35 + 0.15 + (nrows + (nrows-1)*0.1)*0.22
+    figh = 0.35 + 0.15 + (nrows + (nrows - 1) * 0.1) * 0.22
     fig, ax = plt.subplots(nrows=nrows, figsize=(6.4, figh))
     fig.subplots_adjust(top=1-.35/figh, bottom=.15/figh, left=0.2, right=0.99)
 
@@ -69,3 +98,6 @@ def plot_color_gradients(cmap_category, cmap_list):
 
     ax.set(xticks=xtick_locs, xticklabels=xticks)
     ax.get_yaxis().set_visible(False)
+
+    plt.show()
+    return fig, ax