--- a/workshop.py
+++ b/workshop.py
@@ -5,14 +5,14 @@
 """Workshop main flow."""
 
 
-# In[1] Import Packages and Data
+# %% [1] Import Packages and Data
 
 # Third-party imports
 
 # Local imports
 
 
-# In[2] Explore the Data
+# %% [2] Explore the Data
 
 # Read data
 
@@ -22,8 +22,7 @@
 
 # TODO: Print last three rows of DataFrame
 
-
-# In[3] Visualization
+# %% [3] Visualization
 
 # Order rows according to date
 
@@ -35,7 +34,6 @@
 
 # TODO: Plot temperature vs date using only the data from 2006
 
-
 # -----------------------------------------------------------------------------
 
 # Plot temperature and humidity in the same plot
@@ -43,14 +41,14 @@
 # TODO: Plot different combinations of the variables, and for different years
 
 
-# In[4] Data summarization and aggregation
+# %% [4] Data summarization and aggregation
 
 # Weather data by year
 
 # TODO: Create and use a function to average the weather data by month
 
 
-# In[5] Data Analysis and Interpretation
+# %% [5] Data Analysis and Interpretation
 
 # Plot correlations
 
@@ -62,21 +60,19 @@
 #       Contrast them with the weather_correlations dataframe.
 
 
-# In[6] Data Modeling and Prediction
+# %% [6] Data Modeling and Prediction
 
 # Get data subsets for the model
 
 # Run regression
 
-# Get coefficients
+# Print coefficients
 
 
-# In[7] Predictive Model Testing and Evaluation
+# %% [7] Predictive Model Testing and Evaluation
 
 # Plot predicted model with test data.
 
-# TODO: Using the coefficients, predict the temperature for a
-#       given level of humidity using the 'predicted_temperature' function
-#       available in utils.
+# TODO: Using the model, predict the temperature for a given level of humidity
 
 # Evaluate model numerically