This PR avoids changing any core steps or learning content, but implements various fixes, improvements and enhancements over the baseline workshop, including the following:

* Add a number of additional clarifications and explanations to help new Spyder users who many not be familiar with some of the basics of using the IDE or automatically know what/where interface elements are
* Significant copyedits on existing README text to correct, clarify, simplify and improve the language used
* Fix several bugs, including two plots not rendering correctly, column names including spaces/special characters, and a few other minor correctness issues
* Simplify the code slightly to follow best practices and aid understanding, while avoiding any impact to explicitness and pedagogical content
* Greatly improve performance by fixing non-standard datetime format with time zones and ms, speeding up datetime conversation by >1000x (3 OoM), from 60 s to 0.05 s; also make a few other efficiency improvements mostly to utils code; net result is total script runtime from 96 s to 13 s (~10x / 1 OoM reduction), 10% smaller repo and data filesize, and 21% less memory usage
* Add full, Numpydoc style docstrings to all utils functions and fix code lint and style issues 
* Implement various formatting fixes and add .gitattributes and .gitignore

When reviewing, to make it easier to view, comment on and suggest changes I suggest viewing it commit by commit, or all commits following the first one since that one just does some minor formatting and adds per-sentance line breaks making following diffs much easier to read,